Added mapbox map and addressbars

+++ src/main/java/com/example/emmisiontracker/domain/stats/StatEntry.java
@@ -1,5 +1,5 @@
 package com.example.emmisiontracker.domain.stats;
 
-public record StatEntry(String name, float[] data) {
+public record StatEntry(String name, double[] data) {
 
 }
