+++ src/main/java/com/example/emmisiontracker/domain/travel/TravelMethodInfo.java
@@ -2,5 +2,5 @@ package com.example.emmisiontracker.domain.travel;
 
 import com.example.emmisiontracker.constants.TravelMethod;
 
-public record TravelMethodInfo(TravelMethod name, float emissionPerKilometer) {
+public record TravelMethodInfo(TravelMethod name, double emissionPerKilometer) {
 }
